% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{awmc.persian.extent.sp}
\alias{awmc.persian.extent.sp}
\title{AWMC Map of Persian Empire.}
\format{SP object}
\source{
\url{http://awmc.unc.edu}
}
\usage{
awmc.persian.extent.sp
}
\description{
Available from the Ancient World Mapping Center under an CC-BY-NC License
}
\keyword{datasets}

