% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{awmc.senatorial.sp}
\alias{awmc.senatorial.sp}
\title{AWMC Map of Senatorial Provinces of the Roman Empire
Available from the Ancient World Mapping Center under an CC-BY-NC License}
\format{SP object}
\source{
\url{http://awmc.unc.edu}
}
\usage{
awmc.senatorial.sp
}
\description{
AWMC Map of Senatorial Provinces of the Roman Empire
Available from the Ancient World Mapping Center under an CC-BY-NC License
}
\keyword{datasets}

