% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darmc.rats.1to1500.sp}
\alias{darmc.rats.1to1500.sp}
\title{DARMC Rat Remains}
\format{SP object}
\source{
\url{http://darmc.harvard.edu/icb/icb.do?keyword=k40248&pageid=icb.page601659}
}
\usage{
darmc.rats.1to1500.sp
}
\description{
Digital Atlas of Roman and Medieval Civilization data on archaeologically record rat remains. Available under a CC-BY-NC-SA license. See source for details.
}
\details{
The file has a few 0,0 lat/long pairs so it is probably best to remove those before plotting.
}
\keyword{datasets}

