% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darmc.shipwrecks.sp}
\alias{darmc.shipwrecks.sp}
\title{DARMC Shipwrekcs (SP)}
\format{SP Object}
\source{
\url{http://darmc.harvard.edu/icb/icb.do?keyword=k40248&pageid=icb.page601659}
}
\usage{
darmc.shipwrecks.sp
}
\description{
Digital Atlas of Roman and Medieval Civilization data on shipwrecks as an sp object. Available under a CC-BY-NC-SA license. See source for details.
}
\keyword{datasets}

