% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nomisma.greek.mints.sp}
\alias{nomisma.greek.mints.sp}
\title{Nomisma.org: Greek Mints (sp)}
\format{SP Object (needs info on columns)}
\source{
\url{http://nomisma.org/}
}
\usage{
nomisma.greek.mints.sp
}
\description{
Data available under a CC-By license. All mints said by Nomisma to be "part of" Greek numismatics are included. Meaning this is a modern term indicating disciplinary practice, not a description of ancient cultural affinity.
}
\keyword{datasets}

