% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orbis.edited.edges}
\alias{orbis.edited.edges}
\title{Edited Edges from the Stanford Orbid Model of the Roman Mediterranean.}
\format{A data frame (needs info on columns)}
\source{
\url{http://orbis.stanford.edu}
}
\usage{
orbis.edited.edges
}
\description{
Edited Edges from the Stanford Orbid Model of the Roman Mediterranean.
}
\keyword{datasets}

