% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orbis.edited.nodes}
\alias{orbis.edited.nodes}
\title{Edited Nodes from the Stanford Orbid Model of the Roman Mediterranean.}
\format{A data frame (needs info on columns)}
\source{
\url{http://orbis.stanford.edu}
}
\usage{
orbis.edited.nodes
}
\description{
Edited Nodes from the Stanford Orbid Model of the Roman Mediterranean.
}
\keyword{datasets}

