% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orbis.nodes}
\alias{orbis.nodes}
\title{Stanford Orbis Model of Roman Connectivity: Nodes}
\format{A data frame (needs info on columns)}
\source{
\url{http://orbis.stanford.edu/api}
}
\usage{
orbis.nodes
}
\description{
The API is public but does not give an explicit license.
}
\keyword{datasets}

