% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxrep.quarries}
\alias{oxrep.quarries}
\title{Oxford Roman Economy Project: Quarries.}
\format{A data frame (needs info on columns)}
\source{
\url{http://oxrep.classics.ox.ac.uk/databases/stone_quarries_database/}
}
\usage{
oxrep.quarries
}
\description{
Oxford Roman Economy Project: Quarries.
}
\keyword{datasets}

