% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramphs.chrono.grps}
\alias{ramphs.chrono.grps}
\title{Roman Amphitheaters: Chronological Groups.}
\format{A data frame (needs info on columns)}
\source{
\url{http://github.com/sfsheath/roman-amphitheaters}
}
\usage{
ramphs.chrono.grps
}
\description{
Provides start dates and end dates for the chronological groups in the chrono.grp column of the ramphs object. merge() can be used to combine the two objects.
}
\keyword{datasets}

