% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ramphs}
\alias{ramphs}
\title{Roman Amphitheaters.}
\format{A data frame (needs info on columns)}
\source{
\url{http://github.com/sfsheath/roman-amphitheaters}
}
\usage{
ramphs
}
\description{
Data for Roman amphitheaters.
}
\keyword{datasets}

