% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weighted.deciles.R
\name{weighted.deciles}
\alias{weighted.deciles}
\title{Weighted Deciles}
\usage{
weighted.deciles(Qv, Wv, decile.title, decile.xlab, decile.ylab, scale.name)
}
\description{
Weighted Cumulative Cumulative Sum of
}

