% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orbis.edges}
\alias{orbis.edges}
\title{Stanford Orbis Model of Roman Connectivity: Edges}
\format{A data frame (needs info on columns)}
\source{
\url{http://purl.stanford.edu/mn425tz9757}
}
\usage{
orbis.edges
}
\description{
Data available under a CC-By license.
}
\keyword{datasets}

