% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxrep.quarries}
\alias{oxrep.quarries}
\title{Oxford Roman Economy Project: Quarries.}
\format{A data frame (needs info on columns)}
\source{
\url{http://oxrep.classics.ox.ac.uk/databases/stone_quarries_database/}
}
\usage{
oxrep.quarries
}
\description{
First go at merging data from the downloable csv and from the JS map version. Probably not yet entirely correct. The data is downloadable from the OXREP website. I could not find an indication of license or restriction on use. Users may want to confer with the author (Ben Russell) before further distribution. The original data can be cited as Russell, B. J. (2013). Gazetteer of Stone Quarries in the Roman World. Version 1.0.
}
\keyword{datasets}

